utilisateur management

+++ app/vue/components/admin/challenge.config.php
@@ -56,8 +56,8 @@
                 </div>
                 <div class="btn-setting">
                     <a
-                        href="/admin/deleteProjet?id=<?= $projet->getIdProjet() ?>&idChallenge=<?= $challenge->getIdChallenge() ?>"><svg
-                            xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
+                        href="/admin/deleteProjet?id=<?= $projet->getIdProjet() ?>&idChallenge=<?= $challenge->getIdChallenge() ?>" onclick="return confirmAction(event)"><svg
+                            xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" >
                             <path fill="black"
                                 d="M6 19a2 2 0 0 0 2 2h8a2 2 0 0 0 2-2V7H6v12M8 9h8v10H8V9m7.5-5l-1-1h-5l-1 1H5v2h14V4h-3.5Z" />
                         </svg></a>
@@ -98,7 +98,7 @@
                 </div>
                 <div class="btn-setting">
                     <a
-                        href="/admin/deleteRessource?id=<?= $res->getId() ?>&idChallenge=<?= $challenge->getIdChallenge() ?>"><svg
+                        href="/admin/deleteRessource?id=<?= $res->getId() ?>&idChallenge=<?= $challenge->getIdChallenge() ?>" onclick="return confirmAction(event)"><svg
                             xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
                             <path fill="black"
                                 d="M6 19a2 2 0 0 0 2 2h8a2 2 0 0 0 2-2V7H6v12M8 9h8v10H8V9m7.5-5l-1-1h-5l-1 1H5v2h14V4h-3.5Z" />
@@ -116,3 +116,4 @@
     </div>
 
 </div>
+<script src="./vue/components/ConfirmAction.js"></script>
