+++ app/vue/components/admin/manage-user.php
@@ -29,7 +29,7 @@
                     </span>
                 </div>
                 <div class="btn-setting">
-                    <a href="/admin/deleteUser?id=<?= $user->getId() ?>">
+                    <a href="/admin/deleteUser?id=<?= $user->getId() ?>"  onclick="return confirmAction(event)" >
                         <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
                             <path fill="black"
                                 d="M6 19a2 2 0 0 0 2 2h8a2 2 0 0 0 2-2V7H6v12M8 9h8v10H8V9m7.5-5l-1-1h-5l-1 1H5v2h14V4h-3.5Z" />
@@ -48,3 +48,5 @@
     </div>
 
 </div>
+
+<script src="./vue/components/ConfirmAction.js"></script>
